Extract Runner logic into separate class
